fix(momentum): cast retorno_12_1/z_score to native float, fixing checkpoint crash
A live end-to-end smoke test (real Anthropic + Alpaca APIs, checkpoint_enabled=True)
surfaced "Type is not msgpack serializable: numpy.float64" on every attempt, right
after the graph completed. pandas .iloc scalars on a float64 column are
numpy.float64, and round() on a numpy.float64 returns another numpy.float64, not
a native Python float -- so retorno_12_1/z_score stayed numpy-typed all the way
into graph state despite the MomentumMetrics NamedTuple's `float` annotation
(unenforced at runtime). SqliteSaver's msgpack checkpoint serializer doesn't know
numpy.float64, so every real checkpointed run would have failed identically,
silently skipping every ticker every day for the full 4-6 week validation window
(D-08's retry-then-skip masked it as a normal skip, always exiting 0).

Never caught before because unit tests assert value/magnitude, not exact type,
and integration tests don't checkpoint via msgpack -- only a real checkpointed
run against real data hit it.

Fixed by wrapping both round() calls in momentum_calculator.py with float().
Regression test asserts type(...) is float (not isinstance, since numpy.float64
passes isinstance(x, float) but still isn't msgpack-serializable).

+++ tradingagents/agents/utils/momentum_calculator.py
@@ -97,7 +97,13 @@ def compute_momentum(ticker: str, as_of_date: str) -> MomentumMetrics:
             )
 
         retorno_12_1 = (close_1m_ago - close_12m_ago) / close_12m_ago
-        retorno_12_1 = round(retorno_12_1, 6)  # 6 decimals as per D-07
+        # float() is required here: round() on a numpy.float64 (what pandas .iloc
+        # scalars actually are, despite the MomentumMetrics.retorno_12_1: float
+        # annotation) returns another numpy.float64, not a native Python float.
+        # An un-cast numpy.float64 flowing into graph state fails msgpack
+        # serialization at checkpoint time (SqliteSaver), which only surfaces on
+        # a real checkpointed run, not the mocked/dict-state unit tests.
+        retorno_12_1 = float(round(retorno_12_1, 6))  # 6 decimals as per D-07
 
         # Compute z-score from rolling 12-1 distribution (D-15)
         # Build rolling returns: shift(22) to skip month, shift(253) for 12-month lookback
@@ -121,7 +127,7 @@ def compute_momentum(ticker: str, as_of_date: str) -> MomentumMetrics:
                 z_score = 0.0
             else:
                 z_score = (retorno_12_1 - mean) / std
-            z_score = round(z_score, 4)  # 4 decimals as per D-15
+            z_score = float(round(z_score, 4))  # 4 decimals as per D-15; see float() note above
 
         # Derive confidence level from z-score magnitude (D-15)
         confidence_level = derive_confidence_level(z_score)
